added speed to AI and state to AniTimer
AniTimer can now keep track of different states with a dictionary and
can store the frames it needs to switch to based on the state.

AI now has a speed variable which is automatically applied when the
enemy is in idle mode.

added many more changes to make anitimer much more usable for enemies